requirements.txt
certifi==2018.8.24
dj-database-url==0.5.0
Django==2.0.3
django-heroku==0.3.1
django-languages==0.1
django-rest-framework==0.1.0
djangorestframework==3.8.2
findspark==1.1.0
gunicorn==19.9.0
numpy==1.14.0
Pillow==5.2.0
pipenv==2018.7.1
psycopg2==2.7.5
python-decouple==3.1
pytz==2018.3
redis==2.10.6
virtualenv==16.0.0
virtualenv-clone==0.3.0
whitenoise==4.1
